full rework ended

+++ rect.go
@@ -2,18 +2,18 @@ package intermplay
 
 type Rect struct {
 	origin Vector2
-	w, h   int
+	w, h   float32
 }
 
-func (rect *Rect) Size() (int, int) {
+func (rect *Rect) Size() (float32, float32) {
 	return rect.w, rect.h
 }
 
-func (rect *Rect) W() int {
+func (rect *Rect) W() float32 {
 	return rect.w
 }
 
-func (rect *Rect) H() int {
+func (rect *Rect) H() float32 {
 	return rect.h
 }
 
@@ -21,9 +21,9 @@ func (rect *Rect) Origin() Vector2 {
 	return rect.origin
 }
 
-func NewRect(x, y, width, height int) Rect {
+func NewRect(x, y, width, height float32) Rect {
 	return Rect{
-		origin: NewVector2(x, y),
+		origin: NewVector2(float32(x), float32(y)),
 		w:      width,
 		h:      height,
 	}
